clean up firebase stuff, add post preview
