+++ app/api/addKey/route.ts
@@ -1,19 +1,7 @@
 import { NextResponse } from "next/server"
 import * as admin from "firebase-admin"
 
-import "firebase/firestore"
-
-if (!admin.apps.length) {
-  admin.initializeApp({
-    credential: admin.credential.cert({
-      projectId: process.env.NEXT_PUBLIC_PROJECT_ID,
-      clientEmail: process.env.FIREBASE_CLIENT_EMAIL,
-      privateKey: process.env.FIREBASE_PRIVATE_KEY!.replace(/\\n/g, "\n"),
-    }),
-  })
-}
-
-const db = admin.firestore()
+import db from "@/app/lib/firebase"
 
 async function addPubKey(pubkey: string, domain: string, time: Date) {
   // Their firebase proof expiration is 2 weeks from when they last generated the proof.
