+++ app/api/comment/route.ts
@@ -2,19 +2,8 @@ import { randomUUID } from "crypto"
 import { NextResponse } from "next/server"
 import { Post } from "@/types"
 import { ec as EC } from "elliptic"
-import * as admin from "firebase-admin"
 
-if (!admin.apps.length) {
-  admin.initializeApp({
-    credential: admin.credential.cert({
-      projectId: process.env.NEXT_PUBLIC_PROJECT_ID,
-      clientEmail: process.env.FIREBASE_CLIENT_EMAIL,
-      privateKey: process.env.FIREBASE_PRIVATE_KEY!.replace(/\\n/g, "\n"),
-    }),
-  })
-}
-
-const db = admin.firestore()
+import db from "@/app/lib/firebase"
 
 export async function comment(
   postId: string,
