+++ app/lib/firebase.ts
@@ -1,0 +1,17 @@
+import * as admin from "firebase-admin"
+
+import "firebase/firestore"
+
+if (!admin.apps.length) {
+  admin.initializeApp({
+    credential: admin.credential.cert({
+      projectId: process.env.NEXT_PUBLIC_PROJECT_ID,
+      clientEmail: process.env.FIREBASE_CLIENT_EMAIL,
+      privateKey: process.env.FIREBASE_PRIVATE_KEY!.replace(/\\n/g, "\n"),
+    }),
+  })
+}
+
+const db = admin.firestore()
+
+export default db
